% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_marginals.R
\name{bind_marginals}
\alias{bind_marginals}
\title{Joins a copula with two distributions}
\usage{
bind_marginals(copula, dst_x, dst_y)
}
\arguments{
\item{copula}{copula object}

\item{dst_x}{marginal distribution}

\item{dst_y}{marginal distribution}
}
\value{
A list with copula, dst_x and dst_y objects with their names retained.
The list belongs to cop_dst class, which is a class of bi_dst, which is a class of mult_dst.
}
\description{
Creates a joint distribution object from a copula by binding marginal distributions
}
\examples{
bind_marginals(cop_ig(theta = 1, alpha = 3), mtcars$hp, mtcars$disp)
bind_marginals(copula_object, rnorm(50, 2), rnorm(50, 15))

}
