% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_marginals_copula.R
\name{bind_marginals_copula}
\alias{bind_marginals_copula}
\title{Title: bind_marginals_copula
purpose: Transforms the copula distribution from uniform to its scale values such as m3/s}
\usage{
bind_marginals_copula(d1, d2, cop)
}
\arguments{
\item{d1}{\itemize{
\item the first distribution to compare
}}

\item{d2}{\itemize{
\item the second distribution to compare
}}

\item{cop}{\itemize{
\item any copula object of class "bi_cop"
}}
}
\value{
\itemize{
\item cop.d which is a bivariate distribution
}
}
\description{
Title: bind_marginals_copula
purpose: Transforms the copula distribution from uniform to its scale values such as m3/s
}
\examples{
-
}
