% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_cdf.R
\name{eval_cdf}
\alias{eval_cdf}
\alias{enframe_cdf}
\title{Cumulative Distribution Function}
\usage{
eval_cdf(copula, u, v)

enframe_cdf(..., u, v, fn_prefix = "cdf", sep = "_")
}
\arguments{
\item{copula, ...}{A copula, or possibly multiple
copulas in the case of \code{...}.}

\item{u, v}{Vectors of values to evaluate the copula at.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one copula, the
character that will be separating the \code{fn_name} and the copula name.}
}
\value{
The evaluated cdf in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a copula's cumulative distribution function (cdf).
}
\examples{
d1 <- cop_ig(20, 4)
d2 <- cop_igl(1.1)
eval_cdf(d1, u = 0.4, v = 1:9 / 10)
enframe_cdf(d1, u = 0.4, v = 1:9 / 10)
enframe_cdf(d1, d2, u = 1:9 / 10, v = 1:9 / 10)
}
\seealso{
Other distributional representations: 
\code{\link{eval_density}()},
\code{\link{eval_pmf}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
