% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataframe_to_tibble.R
\name{convert_dataframe_to_tibble}
\alias{convert_dataframe_to_tibble}
\title{Convert a data frame to a tibble}
\usage{
convert_dataframe_to_tibble(df)
}
\arguments{
\item{df}{Data frame to possibly convert to a tibble.}
}
\value{
If the user has the tibble package installed, returns the inputted
data frame converted to a tibble. Otherwise, returns the input.
}
\description{
If the tibble package is installed, convert a data frame to a tibble.
}
