% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_ig.R
\name{cop_ig}
\alias{cop_ig}
\alias{cop_igl}
\title{Integrated Gamma Copula}
\usage{
cop_ig(theta, alpha)

cop_igl(alpha)
}
\arguments{
\item{theta, alpha}{Parameters; positive numeric.}
}
\description{
Makes a copula belonging to the Integrated Gamma (IG)
and IG Limit (IGL) families of copulas.
}
\examples{
cop_ig(0.3, 1.4)
cop_igl(1.4)

}
