% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cop_rvine_cop_lib.R
\name{cop_rvine_cop_lib}
\alias{cop_rvine_cop_lib}
\title{Title:Cop_rvine_cop_lib
Purpose: Creates a class object to store the copula parameters}
\usage{
cop_rvine_cop_lib(dist, rotation, parameters)
}
\arguments{
\item{dist}{\itemize{
\item The family name of the copula being used
}}

\item{rotation}{\itemize{
\item the rotation of the copula
}}

\item{theta}{-The theta parameter to scale the copula}
}
\value{
res - a class object for the copula which is a list holding the parameters of the copula
}
\description{
Title:Cop_rvine_cop_lib
Purpose: Creates a class object to store the copula parameters
}
\examples{
-
}
