% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_rvinecopulib.R
\name{cop_rvinecopulib}
\alias{cop_rvinecopulib}
\title{Create a copula from the rvinecopulib package}
\usage{
cop_rvinecopulib(dist, rotation, parameters)
}
\arguments{
\item{dist}{The family name of the copula.}

\item{rotation}{The rotation of the copula; 0, 90, 180, or 270.}

\item{parameters}{A vector of parameters anticipated by rvinecopulib.}
}
\value{
A copula object of class \code{"cop_rvinecopulib"}.
}
\description{
Create a copula from the rvinecopulib package
}
\examples{
cop_rvinecopulib("frank", rotation = 0, parameters = 3)
}
