% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{is_copula}
\alias{is_copula}
\alias{is.copula}
\alias{copula}
\title{Copula Objects}
\usage{
is_copula(object)

is.copula(object)

copula(variable = c("unknown", "continuous", "discrete", "mixed"))
}
\arguments{
\item{object}{Object to be tested}

\item{variable}{Type of random vector: \code{"continuous"}, \code{"discrete"},
\code{"mixed"}, or \code{"unknown"}.}
}
\value{
A copula object with nothing in it.
}
\description{
Currently, this function makes a copula object with nothing in it. The
idea is that you can then set things downstream, with functions such as
set_cdf() and set_mean(). The idea behind this function is expected to be in
flux.
}
