% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_density.R
\name{eval_density}
\alias{eval_density}
\alias{enframe_density}
\title{Probability Density Function}
\usage{
eval_density(copula, u, v, strict = TRUE)

enframe_density(..., u, v, fn_prefix = "density", sep = "_", strict = TRUE)
}
\arguments{
\item{copula, ...}{A copula, or possibly multiple
copulas in the case of \code{...}.}

\item{u, v}{Vectors of values to evaluate the copula at.}

\item{strict}{Only evaluate when the density exists? \code{TRUE} if so;
if \code{FALSE}, evaluates the derivative of the cdf.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one copula, the
character that will be separating the \code{fn_name} and the copula name.}
}
\value{
The evaluated density in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a copula's probability density function (pdf).
}
\examples{
d1 <- cop_ig(20, 4)
d2 <- cop_igl(1.1)
eval_density(d1, u = 0.4, v = 1:9 / 10)
enframe_density(d1, u = 0.4, v = 1:9 / 10)
enframe_density(d1, d2, u = 1:9 / 10, v = 1:9 / 10)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_pmf}()}
}
\concept{distributional representations}
