% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindep.R
\name{dindep}
\alias{dindep}
\title{density of the independence copula}
\usage{
dindep(vector_1, vector_2)
}
\arguments{
\item{vector_1}{A numeric vector}

\item{vector_2}{A numeric vector}
}
\value{
A numeric vector
}
\description{
identifies the density of the independence copula,
1 everywhere on the unit square, and 0 elsewhere
}
\examples{
dindep(c(1, 0.3, 5, 0.8), c(0.5, 0.7, 3, 8))

}
