% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bi_density.bicop_rvinecopulib.R
\name{eval_bi_density.bicop_rvinecopulib}
\alias{eval_bi_density.bicop_rvinecopulib}
\title{Density of an rvinecopulib copula}
\usage{
eval_bi_density.bicop_rvinecopulib(distribution, x, y)
}
\arguments{
\item{distribution}{An rvinecopulib copula, as in the output of
\code{cop_rvinecopulib()}.}

\item{x, y}{Vectors of values between 0 and 1 in which to evaluate the
copula density.}
}
\value{
A vector of copula density values.
}
\description{
Evaluates the density of a copula from the rvinecopulib package.
}
\examples{
cop <- cop_rvinecopulib("frank", rotation = 0, parameters = 2)
eval_bi_density(cop, c(0.2, 0.7, 0.4), 0.2)
}
