% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bi_density.bidst_cop.R
\name{eval_bi_density.bidst_cop}
\alias{eval_bi_density.bidst_cop}
\title{Density of a copula-based bivariate distribution}
\usage{
eval_bi_density.bidst_cop(distribution, x, y)
}
\arguments{
\item{distribution}{Bivariate distribution.}

\item{x, y}{Vectors of values to evaluate the distribution at.}
}
\value{
Vector of density values.
}
\description{
Density of a copula-based bivariate distribution
}
\examples{
d <- bind_marginals(
  distionary::dst_norm(0, 1),
  distionary::dst_exp(1),
  cop_ig(2, 4)
)
eval_bi_density(d, x = -3:3, y = 1)
}
