% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_density_Copula.r
\name{eval_bi_density}
\alias{eval_bi_density}
\title{eval_density}
\usage{
eval_bi_density(dist, x, y)
}
\arguments{
\item{dist}{the distribution of a copula, bivariate or multivariate distribution}

\item{x}{the x variable distribution}

\item{y}{the y variable distribution}
}
\value{
The probabilities of occurence for two points on a copula distribution
}
\description{
eval_density
}
\examples{
- 
}
