% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bi_density_bi_cop_dst.R
\name{eval_bi_density_bi_dst_cop}
\alias{eval_bi_density_bi_dst_cop}
\title{Title: eval_bi_density_bi_cop_dst bivariate distribution from a copula
purpose : evaluate the density of the copula distribution}
\usage{
eval_bi_density_bi_dst_cop(distribution, x, y)
}
\arguments{
\item{x}{\itemize{
\item the first marginal distribution
}}

\item{y}{\itemize{
\item the second marginal distribution
}}
}
\value{
\itemize{
\item density values in the form of a list
}
}
\description{
Title: eval_bi_density_bi_cop_dst bivariate distribution from a copula
purpose : evaluate the density of the copula distribution
}
\examples{
-
}
