% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_bi_pdf_bi_cop_rvinecopulib.R
\name{eval_bi_pdf_bi_cop_rvinecopulib}
\alias{eval_bi_pdf_bi_cop_rvinecopulib}
\title{Title: eval_bi_pdf_bi_cop_rvinecopulib - copula density}
\usage{
eval_bi_pdf_bi_cop_rvinecopulib(distribution, x, y)
}
\arguments{
\item{distribution}{\itemize{
\item output from Cop_rvine_cop_lib (dname, theta, rotation)
}}

\item{x}{\itemize{
\item the first distribution
}}

\item{y}{\itemize{
\item the second distribution
}}

\item{cop}{\itemize{
\item the parameters for the copula that is the output of Cop_rvine_cop_lib
}}
}
\value{
\itemize{
\item a list of values that is the new density distribution for a pdf
}
}
\description{
Title: eval_bi_pdf_bi_cop_rvinecopulib - copula density
}
\examples{
-
}
