% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dindep.R, R/pindep.R
\name{dindep}
\alias{dindep}
\alias{pindep}
\title{The Independence Copula}
\usage{
dindep(u, v)

pindep(u, v)
}
\arguments{
\item{u, v}{Numeric vectors.}
}
\value{
\code{dindep} gives the density, and \code{pindep} gives the cdf, both as
numeric vectors.
}
\description{
Density and cdf of the independence copula.
}
\examples{
dindep(c(1, 0.3, 5, 0.8), c(0.5, 0.7, 3, 8))
dindep(c(0.2, 0.3), 4)
pindep(c(1, 0.2), 0.5)
pindep(c(-2, 8, NA), c(3, 4, 0.1))
pindep(c(-2, 8), c(4, 0.1))
}
