% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copula.R
\name{new_copula}
\alias{new_copula}
\title{Constructor Function for "cop" Objects}
\usage{
new_copula(l, variable, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a copula object.}

\item{variable}{Type of random vector: \code{"continuous"}, \code{"discrete"},
\code{"mixed"}, or \code{"unknown"}.}

\item{...}{Attributes to add to the list.}

\item{class}{If making a subclass, specify its name here.}
}
\description{
Constructor Function for "cop" Objects
}
