% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_pmf.R
\name{eval_pmf}
\alias{eval_pmf}
\alias{enframe_pmf}
\title{Probability Mass Function}
\usage{
eval_pmf(copula, u, v, strict = TRUE)

enframe_pmf(..., u, v, fn_prefix = "pmf", sep = "_", strict = TRUE)
}
\arguments{
\item{copula, ...}{A copula, or possibly multiple
copulas in the case of \code{...}.}

\item{u, v}{Vectors of values to evaluate the copula at.}

\item{strict}{Only evaluate when the pmf exists? \code{TRUE} if so;
if \code{FALSE}, simply evaluates the probability of the specified outcomes.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one copula, the
character that will be separating the \code{fn_name} and the copula name.}
}
\value{
The evaluated probabilities
in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a copula's probability mass function (pmf).
}
\examples{
d1 <- cop_ig(20, 4)
d2 <- cop_igl(1.1)
eval_pmf(d1, u = 0.4, v = 1:9 / 10, strict = FALSE)
enframe_pmf(d1, u = 0.4, v = 1:9 / 10, strict = FALSE)
enframe_pmf(d1, d2, u = 1:9 / 10, v = 1:9 / 10, strict = FALSE)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_density}()},
\code{\link{eval_survival}()}
}
\concept{distributional representations}
