% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realise.R
\name{realise}
\alias{realise}
\alias{realize}
\alias{observe}
\title{Generate a Sample from a Copula}
\usage{
realise(copula, n = 1)

realize(copula, n = 1)

observe(copula, n = 1)
}
\arguments{
\item{copula}{Copula}

\item{n}{Number of observations to generate}
}
\value{
Vector of independent values drawn from the inputted copula
}
\description{
Draw \code{n} independent observations from a copula.
}
\note{
\code{realise()}, \code{realize()}, and \code{observe()} all do the same thing.
Although the language \code{observe()} is more common, it's only an alias here,
because this function name is common amongst other packages.
}
