% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_survival.R
\name{eval_survival}
\alias{eval_survival}
\alias{enframe_survival}
\title{Survival Function}
\usage{
eval_survival(copula, u, v)

enframe_survival(..., u, v, fn_prefix = "survival", sep = "_")
}
\arguments{
\item{copula, ...}{A copula, or possibly multiple
copulas in the case of \code{...}.}

\item{u, v}{Vectors of values to evaluate the copula at.}

\item{fn_prefix}{For \code{enframe_}, name of the function to
appear in the column(s).}

\item{sep}{When \code{enframe}'ing more than one copula, the
character that will be separating the \code{fn_name} and the copula name.}
}
\value{
The evaluated survival function
in vector form (for \code{eval_}) and data frame
or tibble form (for \code{enframe_}).
}
\description{
Access a copula's survival function.
}
\examples{
d1 <- cop_ig(20, 4)
d2 <- cop_igl(1.1)
eval_survival(d1, u = 0.4, v = 1:9 / 10)
enframe_survival(d1, u = 0.4, v = 1:9 / 10)
enframe_survival(d1, d2, u = 1:9 / 10, v = 1:9 / 10)
}
\seealso{
Other distributional representations: 
\code{\link{eval_cdf}()},
\code{\link{eval_density}()},
\code{\link{eval_pmf}()}
}
\concept{distributional representations}
