% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{variable}
\alias{variable}
\title{Type of random vector represented by a copula}
\usage{
variable(copula)
}
\arguments{
\item{copula}{Copula}
}
\value{
As of now, the type of variable that it is -- one
of "continuous", "discrete", or "mixed".
}
\description{
Copulas will either be discrete, continuous, or mixed.
}
