% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_marginals.R
\name{bind_marginals}
\alias{bind_marginals}
\title{Make a multivariate distribution through a copula}
\usage{
bind_marginals(d1, d2, cop)
}
\arguments{
\item{d1, d2}{The first and second marginal distributions to bind.}

\item{cop}{A copula object (has class "bicop").}
}
\value{
A bivariate probability distribution of class \code{"bidst_cop"}
}
\description{
In other words, this function transforms a copula from its uniform scale
to the original scale of the data, such as m3/s.
}
\examples{
d1 <- distionary::dst_norm(0, 1)
d2 <- distionary::dst_exp(1)
bind_marginals(
  d1, d2,
  cop_rvinecopulib("gumbel", rotation = 0, parameters = 2)
)
bind_marginals(
  d1, d2,
  cop_ig(2, 3)
)
}
