% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pindep.R
\name{pindep}
\alias{pindep}
\title{Cdf of the independence copula}
\usage{
pindep(u, v)
}
\arguments{
\item{u}{A numeric vector}

\item{v}{A numeric vector}
}
\value{
A numeric vector
}
\description{
Identifies the cdf of the independence copula:
u * v - if both vectors are located on the unit square;
1 - if both vectors are larger than 1;
0 - if one of the vectors is below 0;
u value - if v is larger than 1;
v value - if u is larger than 1
}
\examples{
pindep(c(1, 0.2), c(0.5))
pindep(c(-2, 8, NA), c(3, 4, 0.1))
pindep(c(-2, 8), c(4, 0.1))

}
